move covariance to under twobody
